backend/requirements.txt
fastapi
uvicorn[standard]
shapely
google-generativeai
python-dotenv
httpx
